0.81.2-0BX01 - Cracklib updated to last version from SVN. Still blows up. Disabled. - i18n PHP module build disabled, as we now replace it with a PHP class.

+++ BlueOnyx/5209R/platform/i18n/cracklib/crack/libcrack/src/rules.c
@@ -433,7 +433,7 @@ cracklib_mangle(char *input, char *control)
 					return (char *) 0;
 				}
 				
-				if (strlen(area) <= limit) {
+				if ((int) strlen(area) <= limit) {
 					return (char *) 0;
 				}
 				break;
@@ -450,7 +450,7 @@ cracklib_mangle(char *input, char *control)
 					return (char *) 0;
 				}
 				
-				if (strlen(area) >= limit) {
+				if ((int) strlen(area) >= limit) {
 					return (char *) 0;
 				}
 				break;
